+++ app/Modules/Module/PermissionsController.php
@@ -50,6 +50,10 @@ class PermissionsController extends ApiController
      */
     public function destroy(Module $module, Permission $permission)
     {
+        if (Module::SYSTEM_MODULE == optional($module->parentModule)->key) {
+            return $this->respondForbidden();
+        }
+
         $permission->delete();
 
         return $this->respondNoContent();
